// IsveskiDemo_NodeJS/clientcode/model/template.ts
/**
 * Isveski Client
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { TemplateTypeEnum } from './templateTypeEnum';

export class Template {
    'templateType': TemplateTypeEnum;

    static discriminator: string | undefined = "templateType";

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "templateType",
            "baseName": "templateType",
            "type": "TemplateTypeEnum"
        }    ];

    static getAttributeTypeMap() {
        return Template.attributeTypeMap;
    }
}

export namespace Template {
}