// IsveskiDemo_NodeJS/servercode/model/requestParameter.ts
/**
 * Isveski Client Server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { RequestTicket } from './requestTicket';

export class RequestParameter {
    'sensorId': string;
    'sensorName': string;
    'clientName': string;
    'userId': string;
    'communicationId': string;
    'name': string;
    'language': string;
    'tickets': Array<RequestTicket>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "sensorId",
            "baseName": "sensorId",
            "type": "string"
        },
        {
            "name": "sensorName",
            "baseName": "sensorName",
            "type": "string"
        },
        {
            "name": "clientName",
            "baseName": "clientName",
            "type": "string"
        },
        {
            "name": "userId",
            "baseName": "userId",
            "type": "string"
        },
        {
            "name": "communicationId",
            "baseName": "communicationId",
            "type": "string"
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "language",
            "baseName": "language",
            "type": "string"
        },
        {
            "name": "tickets",
            "baseName": "tickets",
            "type": "Array<RequestTicket>"
        }    ];

    static getAttributeTypeMap() {
        return RequestParameter.attributeTypeMap;
    }
}